build: split index file by application
This commit splits the React application's `index.ts` into two index
files, `index-breakout.tsx` and `index-sidebar.tsx`. This allows us to
use different SFCC APIs for each "index" file.

This is done by exporting a function that invokes what was previously in
`index.ts` in each file. When invoked, that function will in turn return
all the logic that was previously in `index.ts` but is now in
`index-<my-app>.ts`. This way, the markup for `breakout` and `sidebar`
apps can differ slightly.

When building, `yarn deploy`, the ENV variables are read to decide which
"build" function to use. IE, which "index" file to use.
