+++ frontend/src/index.tsx
@@ -1,38 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { App } from "./components/";
-import "./styles/index.css";
+import { createBreakoutApp } from "./index-breakout";
+import { createSidebarApp } from "./index-sidebar";
 
-declare const listen: Function;
-
-var rootEditorElement;
-/**
- * Initializes the base markup before page is ready. This is not part of the API, and called explicitly at the end of this module.
- */
-function init() {
-  rootEditorElement = document.createElement("div");
-  rootEditorElement.innerHTML = `<h3>Should be replaced by React</h3>`;
-  document.body.appendChild(rootEditorElement);
-  ReactDOM.render(
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>,
-    rootEditorElement
-  );
+if (process.env.REACT_APP_APP_TYPE === "sidebar") {
+  createSidebarApp();
+} else if (process.env.REACT_APP_APP_TYPE === "breakout") {
+  createBreakoutApp();
+} else {
+  throw new Error("REACT_APP_APP_TYPE is not set");
 }
-
-listen("sfcc:ready", () => {
-  init();
-});
-
-// When a value was selected
-listen("sfcc:value", (value: any) => {});
-// When the editor must require the user to select something
-listen("sfcc:required", (value: any) => {});
-// When the editor is asked to disable its controls
-listen("sfcc:disabled", (value: any) => {});
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
